services/api/pyproject.toml
[project]
#About
name = "api"
version = "0.1.0"
description = "A FastAPI app template with basic User capabilities"
authors = [{ name = "Narsonos", email = "narsonos39@gmail.com" } ]

#Dependencies
requires-python = ">=3.12,<4.0"
package-mode = false
dependencies = [
    "fastapi[standard]==0.115.7",
    "httpx==0.28.1",
    "uvicorn==0.34.0",
    "httptools==0.6.4",
    "loguru==0.7.3",
    "tzlocal==5.3",
    "pytz==2025.1",
    "sqlalchemy==2.0.38",
    "sqlmodel==0.0.24",
    "redis==5.2.1",
    "pyjwt==2.10.1",
    "cryptography==44.0.2",
    "gunicorn>=23.0.0",
    "passlib[bcrypt]==1.7.4",
    "pytest-sugar (>=1.1.1,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "prometheus-client (>=0.23.1,<0.24.0)",
    "celery (>=5.5.3,<6.0.0)",
    "python-json-logger (>=3.3.0,<4.0.0)",
    "opentelemetry-api (>=1.37.0,<2.0.0)",
    "opentelemetry-sdk (>=1.37.0,<2.0.0)",
    "opentelemetry-exporter-otlp (>=1.37.0,<2.0.0)",
    "opentelemetry-instrumentation-fastapi (>=0.58b0,<0.59)",
    "opentelemetry-instrumentation-logging (>=0.58b0,<0.59)",
    "opentelemetry-instrumentation-wsgi (>=0.58b0,<0.59)",
    "opentelemetry-exporter-prometheus (>=0.58b0,<0.59)",
    "opentelemetry-instrumentation (>=0.58b0,<0.59)",
    "opentelemetry-instrumentation-redis (>=0.58b0,<0.59)",
    "opentelemetry-instrumentation-sqlalchemy (>=0.58b0,<0.59)",
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["./tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-ra -v --cov=app --cov-report=term-missing"
markers = ['exceptions', 'services', 'models']
asyncio_mode = 'strict'
asyncio_default_test_loop_scope = 'session'
asyncio_default_fixture_loop_scope = "session"

[tool.coverage.run]
omit = [
    "tests/*",
    "app/common/*",
    "app/main.py",
    "app/gunicorn_conf.py",
    "*/__init__.py",
    "*/interfaces/*",
    "*/dependencies.py",
    "app/presentation/exception_handlers.py"
]

[tool.coverage.report]
fail_under = 95

[tool.pyright]
exclude = [".venv", "**/__pycache__"]
venvPath = "."                          
venv = ".venv"